Create RelatedProducts shelf and a ShelfContainer
