+++ react/locales/en-US.json
@@ -1,18 +1,19 @@
 {
-    "editor.shelf.title": "Shelf",
-    "editor.shelf.description": "A product shelf featuring a collection",
-    "editor.shelf.category.title": "Category",
-    "editor.shelf.collection.title": "Collection",
-    "editor.shelf.orderBy.title": "List Ordenation",
-    "editor.shelf.maxItems.title": "Max Items",
-    "editor.shelf.itemsPerPage.title": "Items Per Page",
-    "editor.shelf.scrollType.title": "Scroll Type",
-    "editor.shelf.arrows.title": "Arrows",
-    "editor.shelf.titleText.title": "Title Text",
-    "editor.shelf.summary.title": "Product Summary",
-    "editor.shelf.orderType.sales": "Sales",
-    "editor.shelf.orderType.priceDesc": "Price, descending",
-    "editor.shelf.orderType.priceAsc": "Price, ascending",
-    "editor.shelf.scrollType.byPage": "By Page",
-    "editor.shelf.scrollType.oneByOne": "One by One"
+  "editor.shelf.title": "Shelf",
+  "editor.shelf.description": "A product shelf featuring a collection",
+  "editor.shelf.category.title": "Category",
+  "editor.shelf.collection.title": "Collection",
+  "editor.shelf.orderBy.title": "List Ordenation",
+  "editor.shelf.maxItems.title": "Max Items",
+  "editor.shelf.itemsPerPage.title": "Items Per Page",
+  "editor.shelf.scrollType.title": "Scroll Type",
+  "editor.shelf.arrows.title": "Arrows",
+  "editor.shelf.titleText.title": "Title Text",
+  "editor.shelf.summary.title": "Product Summary",
+  "editor.shelf.orderType.sales": "Sales",
+  "editor.shelf.orderType.priceDesc": "Price, descending",
+  "editor.shelf.orderType.priceAsc": "Price, ascending",
+  "editor.shelf.scrollType.byPage": "By Page",
+  "editor.shelf.scrollType.oneByOne": "One by One",
+  "editor.relatedProducts.titleText.default": "Related Products"
 }
