+++ react/locales/es-AR.json
@@ -1,18 +1,19 @@
 {
-    "editor.shelf.title": "Vitrina",
-    "editor.shelf.description": "Una vitrina con una colección de productos",
-    "editor.shelf.category.title": "Categoría",
-    "editor.shelf.collection.title": "Colección",
-    "editor.shelf.orderBy.title": "Ordenación de la Lista",
-    "editor.shelf.maxItems.title": "Cantidad máxima de Elementos",
-    "editor.shelf.itemsPerPage.title": "Productos por Página",
-    "editor.shelf.scrollType.title": "Tipo de Desplazamiento",
-    "editor.shelf.arrows.title": "Dardos",
-    "editor.shelf.titleText.title": "Texto del Título",
-    "editor.shelf.summary.title": "Resumen del Producto",
-    "editor.shelf.orderType.sales": "Ventas",
-    "editor.shelf.orderType.priceDesc": "Precio, descendiendo",
-    "editor.shelf.orderType.priceAsc": "Precio, ascendente",
-    "editor.shelf.scrollType.byPage": "Por Página",
-    "editor.shelf.scrollType.oneByOne": "Uno a Uno"
+  "editor.shelf.title": "Vitrina",
+  "editor.shelf.description": "Una vitrina con una colección de productos",
+  "editor.shelf.category.title": "Categoría",
+  "editor.shelf.collection.title": "Colección",
+  "editor.shelf.orderBy.title": "Ordenación de la Lista",
+  "editor.shelf.maxItems.title": "Cantidad máxima de Elementos",
+  "editor.shelf.itemsPerPage.title": "Productos por Página",
+  "editor.shelf.scrollType.title": "Tipo de Desplazamiento",
+  "editor.shelf.arrows.title": "Dardos",
+  "editor.shelf.titleText.title": "Texto del Título",
+  "editor.shelf.summary.title": "Resumen del Producto",
+  "editor.shelf.orderType.sales": "Ventas",
+  "editor.shelf.orderType.priceDesc": "Precio, descendiendo",
+  "editor.shelf.orderType.priceAsc": "Precio, ascendente",
+  "editor.shelf.scrollType.byPage": "Por Página",
+  "editor.shelf.scrollType.oneByOne": "Uno a Uno",
+  "editor.relatedProducts.titleText.default": "Productos relacionados"
 }
