+++ react/locales/pt-BR.json
@@ -1,18 +1,19 @@
 {
-    "editor.shelf.title": "Vitrine",
-    "editor.shelf.description": "Uma vitrine com uma coleção de produtos",
-    "editor.shelf.category.title": "Categoria",
-    "editor.shelf.collection.title": "Coleção",
-    "editor.shelf.orderBy.title": "Ordenação da Lista",
-    "editor.shelf.maxItems.title": "Quantidade máxima de itens",
-    "editor.shelf.itemsPerPage.title": "Itens por Página",
-    "editor.shelf.scrollType.title": "Tipo de Rolagem",
-    "editor.shelf.arrows.title": "Setas",
-    "editor.shelf.titleText.title": "Texto do Título",
-    "editor.shelf.summary.title": "Resumo do Produto",
-    "editor.shelf.orderType.sales": "Vendas",
-    "editor.shelf.orderType.priceDesc": "Preço, descrescente",
-    "editor.shelf.orderType.priceAsc": "Price, crescente",
-    "editor.shelf.scrollType.byPage": "Por Página",
-    "editor.shelf.scrollType.oneByOne": "Um por Um"
+  "editor.shelf.title": "Vitrine",
+  "editor.shelf.description": "Uma vitrine com uma coleção de produtos",
+  "editor.shelf.category.title": "Categoria",
+  "editor.shelf.collection.title": "Coleção",
+  "editor.shelf.orderBy.title": "Ordenação da Lista",
+  "editor.shelf.maxItems.title": "Quantidade máxima de itens",
+  "editor.shelf.itemsPerPage.title": "Itens por Página",
+  "editor.shelf.scrollType.title": "Tipo de Rolagem",
+  "editor.shelf.arrows.title": "Setas",
+  "editor.shelf.titleText.title": "Texto do Título",
+  "editor.shelf.summary.title": "Resumo do Produto",
+  "editor.shelf.orderType.sales": "Vendas",
+  "editor.shelf.orderType.priceDesc": "Preço, descrescente",
+  "editor.shelf.orderType.priceAsc": "Price, crescente",
+  "editor.shelf.scrollType.byPage": "Por Página",
+  "editor.shelf.scrollType.oneByOne": "Um por Um",
+  "editor.relatedProducts.titleText.default": "Produtos relacionados"
 }
