+++ react/queries/relatedProductsQuery.gql
@@ -1,0 +1,39 @@
+query RelatedProducts($slug: String) {
+  product(slug: $slug) {
+    recommendations {
+      view {
+        productId
+        productName
+        description
+        categories
+        link
+        linkText
+        brand
+        items {
+          name
+          itemId
+          referenceId {
+            Value
+          }
+          images {
+            imageUrl
+            imageTag
+          }
+          sellers {
+            commertialOffer {
+              Installments {
+                Value
+                InterestRate
+                TotalValuePlusInterestRate
+                NumberOfInstallments
+                Name
+              }
+              Price
+              ListPrice
+            }
+          }
+        }
+      }
+    }
+  }
+}
